refactor: handler 패키지 통합
