+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/AnnotationHandlerAdapter.java
@@ -1,7 +1,6 @@
-package com.interface21.webmvc.servlet.mvc.handleradapter;
+package com.interface21.webmvc.servlet.mvc.handler;
 
 import com.interface21.webmvc.servlet.ModelAndView;
-import com.interface21.webmvc.servlet.mvc.handlermapping.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
