+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/HandlerAdapterRegistry.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.handleradapter;
+package com.interface21.webmvc.servlet.mvc.handler;
 
 import java.util.List;
