+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/HandlerKey.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.handlermapping;
+package com.interface21.webmvc.servlet.mvc.handler;
 
 import com.interface21.web.bind.annotation.RequestMethod;
