+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/handler/HandlerMappingRegistry.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.handlermapping;
+package com.interface21.webmvc.servlet.mvc.handler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import java.util.List;
