+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/AnnotationHandlerMappingTest.java
@@ -1,7 +1,7 @@
 package com.interface21.webmvc.servlet.mvc;
 
-import com.interface21.webmvc.servlet.mvc.handlermapping.AnnotationHandlerMapping;
-import com.interface21.webmvc.servlet.mvc.handlermapping.HandlerExecution;
+import com.interface21.webmvc.servlet.mvc.handler.AnnotationHandlerMapping;
+import com.interface21.webmvc.servlet.mvc.handler.HandlerExecution;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.junit.jupiter.api.BeforeEach;
